# main-service/docs/swagger.yaml
basePath: /main/api/v1
definitions:
  handlers.ErrorResponse:
    properties:
      message:
        type: string
    type: object
  handlers.SigninRequest:
    properties:
      email:
        example: admin@example.com
        type: string
      password:
        example: admin123
        type: string
    type: object
  handlers.SigninResponse:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      expires_in:
        example: 3600
        type: integer
      token_type:
        example: Bearer
        type: string
    type: object
  handlers.UserResponse:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Main service provides authentication and user management endpoints.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Main Service API
  version: "1.0"
paths:
  /auth/signin:
    post:
      consumes:
      - application/json
      description: Authenticate with email and password to receive a JWT Bearer token
      parameters:
      - description: Credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.SigninRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SigninResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Sign in
      tags:
      - auth
  /me:
    get:
      consumes:
      - application/json
      description: Returns the authenticated user's profile based on the JWT token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - Bearer: []
      summary: Get current user
      tags:
      - user
  /users:
    get:
      consumes:
      - application/json
      description: Returns all registered users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.UserResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - Bearer: []
      summary: List users
      tags:
      - user
securityDefinitions:
  Bearer:
    description: Enter your token (with or without "Bearer " prefix)
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"